Migrated StoreKit/Consumables to unified api
Some int/nint switching and casting

+++ StoreKit/Consumables/CreditManager.cs
@@ -1,5 +1,5 @@
 using System;
-using MonoTouch.Foundation;
+using Foundation;
 
 namespace Consumables {
 	// WARNING: this is a trivial example of tracking a
@@ -14,7 +14,7 @@ namespace Consumables {
 	// NOT how you should build a credits system for iOS.
 	public static class CreditManager {
 		static NSString defaultKey = new NSString("monkeyDollarsBalance");
-		static int monkeyCredits = 0;
+		static nint monkeyCredits = 0;
 		static CreditManager ()
 		{
 			monkeyCredits = NSUserDefaults.StandardUserDefaults.IntForKey(defaultKey);
@@ -22,7 +22,7 @@ namespace Consumables {
 		
 		public static int Balance() {
 			NSUserDefaults.StandardUserDefaults.Synchronize ();
-			return NSUserDefaults.StandardUserDefaults.IntForKey(defaultKey);
+			return (int)NSUserDefaults.StandardUserDefaults.IntForKey(defaultKey);
 		}
 		public static void Add (int moreDollars) {
 			monkeyCredits += moreDollars;
